WS-114: Moves and deletes
