--- src/app/legacy/containers/OptimizelyPageViewTracking/index.jsx
@@ -1,32 +1,0 @@
-import { useState, useContext, useEffect } from 'react';
-import { OptimizelyContext } from '@optimizely/react-sdk';
-import { RequestContext } from '#contexts/RequestContext';
-import useOptimizelyScrollDepth from '#hooks/useOptimizelyScrollDepth';
-import useOptimizelyVariation from '#hooks/useOptimizelyVariation';
-import OPTIMIZELY_CONFIG from '#lib/config/optimizely';
-
-const OptimizelyPageViewTracking = () => {
-  const { isAmp } = useContext(RequestContext);
-  const { optimizely } = useContext(OptimizelyContext);
-  const [pageViewSent, setPageViewSent] = useState(false);
-
-  const experimentVariation = useOptimizelyVariation(OPTIMIZELY_CONFIG.flagKey);
-  const hasVariationKey = experimentVariation !== null;
-
-  const sendPageViewEvent = hasVariationKey && !isAmp && !pageViewSent;
-
-  useOptimizelyScrollDepth();
-
-  useEffect(() => {
-    if (sendPageViewEvent) {
-      optimizely?.onReady().then(() => {
-        optimizely.track('page-views');
-        setPageViewSent(true);
-      });
-    }
-  }, [sendPageViewEvent, optimizely]);
-
-  return null;
-};
-
-export default OptimizelyPageViewTracking;
